[Feat] : Repo Entity, GitHubSearchResponseDTO 생성
Ref: #3
Related to: #8
