+++ GitHubSearch/Data/Repository/GitHubSearchRepository.swift
@@ -1,0 +1,8 @@
+//
+//  GitHubSearchRepository.swift
+//  GitHubSearch
+//
+//  Created by chmini on 2022/07/24.
+//
+
+import Foundation
