+++ GitHubSearch/Domain/Entity/Repo.swift
@@ -1,0 +1,14 @@
+//
+//  Repo.swift
+//  GitHubSearch
+//
+//  Created by chmini on 2022/07/24.
+//
+
+import Foundation
+
+struct Repo {
+    let name: String
+    let fullName: String
+    let starCount: Int
+}
